# openapi.yaml
openapi: 3.1.0
info:
  title: Book For Me MCP Server
  version: 1.0.0
  description: Make AI-powered restaurant calls and receive booking updates.
servers:
  - url: https://bookforme-mcp-server.onrender.com
paths:
  /call:
    post:
      operationId: startCall
      summary: Start a restaurant call
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  description: First name of the user
                lastName:
                  type: string
                  description: Last name of the user
                partyNum:
                  type: integer
                  description: Number of people to book for
                hour:
                  type: integer
                  description: Hour of the reservation (0-23)
                minute:
                  type: integer
                  description: Minute of the reservation (0-59)
                month:
                  type: integer
                  description: Month of the reservation (1-12)
                  minimum: 1
                  maximum: 12
                date:
                  type: integer
                  description: Date of the reservation (1-31)
                  minimum: 1
                  maximum: 31
                language:
                  type: string
                  description: Language to make the call in
                  enum: [en, ja]
                restaurantPhone:
                  type: string
                  description: Phone number of the restaurant to call in international format (e.g. +1234567890)
                userPhone:
                  type: string
                  description: Optional. Phone number of the user (optional)
                restaurantName:
                  type: string
                  description: Optional. Name of the restaurant to call
                custom_request:
                  type: string
                  description: Optional. User's custom requests for the reservation (e.g. "window seat", "high chair", "vegetarian")
              required: [firstName, lastName, partyNum, hour, minute, month, date, language, restaurantPhone]
      responses:
        '200':
          description: Call started
          content:
            application/json:
              schema:
                type: object
                properties:
                  call_id:
                    type: string

  /call-status/{call_id}:
    get:
      operationId: getCallStatus
      summary: Get current status of the restaurant call
      parameters:
        - name: call_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Current call status and conversation log
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: Current call status (`inProgress`, `done`, `failed`)
                  conversation:
                    type: array
                    description: A running log of what has been said on the call so far.
                    items:
                      type: object
                      properties:
                        sender:
                          type: string
                          enum: [agent, receiver]
                        content:
                          type: string

  /call-summary/{call_id}:
    get:
      operationId: getCallSummary
      summary: Get the call summary once complete
      parameters:
        - name: call_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Call summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  isReservationSuccessful:
                    type: boolean
                    description: Whether the reservation was successful
                  summary:
                    type: string
                  transcript:
                    type: array
                    description: A transcript of the entire call